Support calls to list, add xrpks
There's a bit in here that calls for transforms as strings (json)
instead of the Transform object. We may want to pass the Transform in
and do the json conversion in the exporter instead of the app, but it
just felt like the quickest path at the time, esp. since I didn't have
an object whose transform that I wanted to use, and Transforms seem
difficult to use independently og GameObjects. (On the other hand, since
I'm spawning in front of the follower cam, I could've parented an empty
object to that at an offset and used the Transform, but that's one of
those good thoughts that came along later.)