Handle JSONification of transform elements, invert z for web
Previously, the Unity app would be responsible converting the
pos/rot/scale to a JSON string representing the javascript transform.
Now, we take in those components natively and do the conversion in the
exporter. (Still not doing a transform because a transform is bound to a
game object, and we may want to spawn things without a GO sometimes,
such as we're currently doing.)

We now also negate the z internally so that the Unity app can pretend
it's in web space already.